requirements.txt
Django==3.2.3
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-filter==2.4.0
coverage==7.2.5


